requirements.txt
torch
torchvision
segmentation-models-pytorch
timm
Pillow
numpy
tqdm
tifffile